Make examples non-blocking
The README.md examples contain steps that are blocking due to
terraform asking for input. Due to stdin and stderr captured, it
can be hard to debug what is going wrong. The -input=false parameter
to the terraform init, terraform plan and terraform apply calls will
produce an error when required input is not given.