Remove wrapper package.json (#302)
Reference: https://github.com/hashicorp/setup-terraform/issues/290

The `wrapper` directory contains the JavaScript wrapper for the Terraform binary. It currently contains a `package.json` and lock file, however this code is not packaged and released separately. Having the separate package makes it unnecessarily difficult to ensure the subdirectory is built correctly and dependencies are up to date.

This change removes the `wrapper` package files and consolidates the test, lint, and build logic into the main `package.json`.